+++ changelog.md
@@ -1,2 +1,6 @@
-# 1.0.0
+# v1.0.1
+- Added more to the mod description
+- Added refresh button
+- Added an error if server is down
+# v1.0.0
 - Mod Release